+++ Sources/Controllers/ALKConversationViewController+TableView.swift
@@ -544,7 +544,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                 }
                 return cell
             }
-        case .initialEncryptionMessage:
+        case .staticTopMessage:
             let cell = tableView.dequeueReusableCell(forIndexPath: indexPath) as KMStaticTopMessageCell
             cell.updateMessage(viewModel: message)
             return cell