+++ Sources/Utilities/ALMessage+Extension.swift
@@ -137,7 +137,7 @@ extension ALMessage: ALKChatViewModelProtocol {
             return "Document"
         case .allButtons:
             return isMessageEmpty ? "Buttons" : message
-        case .initialEncryptionMessage:
+        case .staticTopMessage:
             return message
         }
     }
@@ -238,7 +238,7 @@ extension ALMessage {
         case ALMESSAGE_CONTENT_TEXT_HTML:
             return richMessageType()
         case ALMESSAGE_CONTENT_INITIAL_STATIC_MESSAGE:
-            return .initialEncryptionMessage
+            return .staticTopMessage
         default:
             guard let attachmentType = getAttachmentType() else { return .text }
             return attachmentType